mptcp: widen offset field in mptcp_data_frag to u32
The offset field in struct mptcp_data_frag is declared as u16, which
can only represent values up to 65535. On architectures with page
sizes larger than 64KB (such as 256KB on PowerPC and Hexagon), the
offset calculation in mptcp_carve_data_frag() can exceed this limit,
causing the u16 field to wrap around.

When offset wraps, mptcp_sendmsg() calculates the destination address
using the wrapped value:

	offset = dfrag->offset + dfrag->data_len;
	page_address(dfrag->page) + offset

This could cause the function to write new network payload into the
start of the page fragment, potentially overwriting the previously
allocated struct mptcp_data_frag and corrupting kernel memory.

Widen the offset field from u16 to u32 to support page sizes up to
4GB. Also change the corresponding local variable in
mptcp_sendmsg_frag() from int to size_t to prevent overflow when
adding offset and data_len.

This is a correctness fix for architectures with 256KB pages and
has no effect on systems with smaller page sizes.

Message-ID: <8b39801c46b80523ee71dfab667117030d2ccdfc.1784188064.git.tanggeliang@kylinos.cn>

+++ net/mptcp/protocol.c
@@ -1373,7 +1373,7 @@ static int mptcp_sendmsg_frag(struct sock *sk, struct sock *ssk,
 			      struct mptcp_sendmsg_info *info)
 {
 	u64 data_seq = dfrag->data_seq + info->sent;
-	int offset = dfrag->offset + info->sent;
+	size_t offset = dfrag->offset + info->sent;
 	struct mptcp_sock *msk = mptcp_sk(sk);
 	struct tcp_sock *tp = tcp_sk(ssk);
 	bool zero_window_probe = false;
